removed dustgrain transfers and added an example file with wcdm and massive neutrinos
