--- reactions/examples/dustgrain/0.15ev/F5_nu015_params.ini
@@ -1,89 +1,0 @@
-output_root = F5_nu015
-get_scalar_cls = F
-get_vector_cls = F
-get_tensor_cls = F
-CMB_outputscale = 7.42835025e12
-get_transfer = T
-accuracy_boost = 3
-l_accuracy_boost = 1
-high_accuracy_default = T
-do_nonlinear = 0
-w = -1
-cs2_lam = 1
-hubble = 67.31
-use_physical = T
-ombh2 = 0.0218
-omch2 = 0.118598
-omnuh2 = 0.00162197
-omk = 0
-MG_flag = 3
-GRtrans = 0.001d0
-QSA_flag = 4
-F_R0 = 0.00001d0
-FRn = 1.d0
-DE_model =            0
-temp_cmb = 2.7255
-helium_fraction = 0.24
-massless_neutrinos = 2.046
-nu_mass_eigenstates = 1
-massive_neutrinos = 1
-share_delta_neff = T
-nu_mass_fractions = 1
-transfer_high_precision = T
-accurate_massive_neutrino_transfers =  F
-transfer_kmax = 100
-transfer_k_per_logint = 0
-transfer_num_redshifts = 3
-transfer_interp_matterpower = F
-transfer_power_var = 7
-transfer_redshift(1) = 1
-transfer_filename(1) = transfer_out_z1.dat
-transfer_matterpower(1) = matterpower_z1.dat
-transfer_redshift(2) = 0.5
-transfer_filename(2) = transfer_out_z0.5.dat
-transfer_matterpower(2) = matterpower_z0.5.dat
-transfer_redshift(3) = 0
-transfer_filename(3) = transfer_out_z0.dat
-transfer_matterpower(3) = matterpower_z0.dat
-DebugParam = 0.0000000000000000
-Alens = 1.0000000000000000
-reionization = T
-re_use_optical_depth = T
-re_optical_depth = 0.09
-re_delta_redshift = 1.5
-re_ionization_frac = -1
-re_helium_redshift = 3.5
-re_helium_delta_redshift = 0.5
-re_helium_redshiftstart = 5.0000000000000000
-pivot_scalar = 0.05
-pivot_tensor = 0.05
-initial_power_num = 1
-scalar_spectral_index(1) = 0.9658
-scalar_nrun(1) = 0
-scalar_nrunrun(1) = 0
-scalar_amp(1) = 2.199e-9
-RECFAST_fudge_He = 0.86
-RECFAST_Heswitch = 6
-RECFAST_Hswitch = T
-RECFAST_fudge = 1.14
-AGauss1 = -0.14000000000000001
-AGauss2 = 7.9000000000000001E-002
-zGauss1 = 7.2800000000000002
-zGauss2 = 6.7300000000000004
-wGauss1 = 0.17999999999999999
-wGauss2 = 0.33000000000000002
-do_lensing_bispectrum = F
-do_primordial_bispectrum = F
-initial_condition = 1
-accurate_polarization = T
-accurate_reionization = T
-accurate_BB = F
-derived_parameters = T
-version_check = Nov16
-do_late_rad_truncation = T
-feedback_level = 1
-output_file_headers = T
-massive_nu_approx = 1
-number_of_threads = 0
-use_spline_template = T
-l_sample_boost = 1
