--- reactions/examples/dustgrain/LCDM_power_all_033.txt
@@ -1,53 +1,0 @@
-    1.16367E-02    1.89019E-03    2.36781E+04
-    1.82539E-02    8.00131E-03    2.59671E+04
-    2.05611E-02    1.29494E-02    2.94063E+04
-    2.46437E-02    1.91652E-02    2.52770E+04
-    2.70496E-02    2.61339E-02    2.60647E+04
-    3.06266E-02    3.50164E-02    2.40605E+04
-    3.55856E-02    4.66438E-02    2.04315E+04
-    4.07105E-02    6.19752E-02    1.81313E+04
-    4.63332E-02    8.21412E-02    1.63010E+04
-    5.30061E-02    1.08385E-01    1.43656E+04
-    6.06399E-02    1.40962E-01    1.24784E+04
-    6.97331E-02    1.81139E-01    1.05445E+04
-    7.97760E-02    2.30522E-01    8.96241E+03
-    9.07941E-02    2.90536E-01    7.66227E+03
-    1.03870E-01    3.64395E-01    6.41847E+03
-    1.18829E-01    4.56795E-01    5.37383E+03
-    1.35943E-01    5.72568E-01    4.49869E+03
-    1.55521E-01    7.17233E-01    3.76377E+03
-    1.77919E-01    9.00095E-01    3.15465E+03
-    2.03543E-01    1.13297E+00    2.65204E+03
-    2.32857E-01    1.43496E+00    2.24338E+03
-    2.66393E-01    1.82682E+00    1.90746E+03
-    3.04758E-01    2.34160E+00    1.63297E+03
-    3.48649E-01    3.02194E+00    1.40750E+03
-    3.98861E-01    3.93073E+00    1.22275E+03
-    4.56304E-01    5.13775E+00    1.06743E+03
-    5.22020E-01    6.74688E+00    9.36206E+02
-    5.97201E-01    8.87612E+00    8.22605E+02
-    6.83209E-01    1.16713E+01    7.22419E+02
-    7.81604E-01    1.52992E+01    6.32466E+02
-    8.94169E-01    1.99687E+01    5.51343E+02
-    1.02295E+00    2.58867E+01    4.77357E+02
-    1.17027E+00    3.32849E+01    4.09939E+02
-    1.33881E+00    4.23900E+01    3.48688E+02
-    1.53162E+00    5.34108E+01    2.93431E+02
-    1.75221E+00    6.65356E+01    2.44132E+02
-    2.00456E+00    8.19750E+01    2.00888E+02
-    2.29325E+00    9.99513E+01    1.63593E+02
-    2.62352E+00    1.20748E+02    1.31995E+02
-    3.00135E+00    1.44870E+02    1.05769E+02
-    3.43360E+00    1.73665E+02    8.46823E+01
-    3.92811E+00    2.05782E+02    6.70173E+01
-    4.49383E+00    2.41380E+02    5.25026E+01
-    5.14102E+00    2.80582E+02    4.07607E+01
-    5.88142E+00    3.23070E+02    3.13459E+01
-    6.52275E+00    3.67020E+02    2.61053E+01
-    7.46215E+00    4.16125E+02    1.97680E+01
-    8.53684E+00    4.71275E+02    1.49525E+01
-    9.76630E+00    5.33645E+02    1.13082E+01
-    1.11728E+01    6.02126E+02    8.52180E+00
-    1.27819E+01    6.82223E+02    6.44867E+00
-    1.46228E+01    7.54949E+02    4.76602E+00
-    1.67287E+01    8.61408E+02    3.63205E+00
