+++ reactions/examples/transfers/baha/baha_lcdm_params.ini
@@ -1,0 +1,92 @@
+highL_unlensed_cl_template = /anaconda3/lib/python3.7/site-packages/camb/HighLExtrapTemplate_lenspotentialCls.dat
+number_of_threads = 0
+DebugParam = 0.0000000000000000
+feedback_level = 1
+get_scalar_cls = F
+get_vector_cls = F
+get_tensor_cls = F
+want_CMB = T
+limber_windows = T
+limber_phiphi = 100
+Do21cm = F
+get_transfer = T
+accuracy_boost = 3
+l_accuracy_boost = 1
+do_nonlinear = 0
+evolve_baryon_cs = F
+evolve_delta_xe = F
+evolve_delta_ts = F
+l_min = 2
+dark_energy_model = fluid
+use_tabulated_w = F
+w = -1.0
+wa = 0.0
+cs2_lam = 1
+hubble = 70.0
+use_physical = T
+ombh2 = 0.022687
+omch2 = 0.111573
+omnuh2 = 0.002597
+omk = 0
+temp_cmb = 2.725
+helium_fraction = 0.24
+massless_neutrinos = 2.046
+nu_mass_eigenstates = 1
+massive_neutrinos = 1
+share_delta_neff = T
+nu_mass_fractions = 1
+transfer_high_precision = T
+accurate_massive_neutrino_transfers = T
+transfer_kmax = 100
+transfer_k_per_logint = 0
+transfer_num_redshifts = 2
+transfer_interp_matterpower = F
+transfer_power_var = 7
+transfer_redshift(1) = 1
+transfer_redshift(2) = 0
+Alens = 1.0000000000000000
+reionization = T
+re_use_optical_depth = T
+re_optical_depth = 0.09
+re_delta_redshift = 1.5
+re_ionization_frac = -1
+re_helium_redshift = 3.5000000000000000
+re_helium_delta_redshift = 0.50000000000000000
+re_helium_redshiftstart = 5.0000000000000000
+pivot_scalar = 0.002
+pivot_tensor = 2e-3
+initial_power_num = 1
+scalar_spectral_index(1) = 0.972
+scalar_nrun(1) = 0
+scalar_nrunrun(1) = 0.0000000000000000
+scalar_amp(1) = 2.45e-09
+recombination_model = Recfast
+RECFAST_fudge_He = 0.86
+RECFAST_Heswitch = 6
+RECFAST_Hswitch = T
+RECFAST_fudge = 1.14
+AGauss1 = -0.14000000000000001
+AGauss2 = 7.9000000000000001E-002
+zGauss1 = 7.2800000000000002
+zGauss2 = 6.7300000000000004
+wGauss1 = 0.17999999999999999
+wGauss2 = 0.33000000000000002
+initial_condition = 1
+use_cl_spline_template = T
+derived_parameters = T
+accurate_polarization = T
+accurate_reionization = T
+accurate_BB = F
+do_late_rad_truncation = T
+massive_nu_approx = 1
+l_sample_boost = 1
+output_root = bacco_wcdm
+transfer_filename(1) = transfer_1.dat
+transfer_matterpower(1) = matterpow_1.dat
+transfer_filename(2) = transfer_0.dat
+transfer_matterpower(2) = matterpow_0.dat
+do_lensing_bispectrum = F
+do_primordial_bispectrum = F
+CMB_outputscale = 7.4311e12
+version_check = Aug18
+output_file_headers = T
